requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
aiofiles
jinja2

